Terminal interface added

+++ kernel/memory/kmalloc.h
@@ -1,4 +1,4 @@
 void kinit_memory();
 void* kmalloc (int size);
-void kmemset (void *addr, int c, size_t n);
-void* kcalloc (int nelem, size_t bytes);
+void kmemset (void *addr, int c, int n);
+void* kcalloc (int nelem, int bytes);
